# api/docs/swagger.yaml
definitions:
  responses.GradeData:
    properties:
      _id:
        type: string
      data:
        items:
          properties:
            grade_distribution: {}
            type:
              type: string
          type: object
        type: array
    type: object
  responses.GradeResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  responses.SectionGradeResponse:
    properties:
      grade_data:
        items:
          $ref: '#/definitions/responses.GradeData'
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
...
host: api.utdnebula.com
info:
  contact: {}
  description: The developer Nebula Labs API for access to pertinent UT Dallas data
  title: dev-nebula-api
  version: 1.0.0
paths:
...
  /course/{id}/grades:
    get:
      description: '"Returns the overall grade distribution for a course"'
      operationId: GradesByCourseID
      parameters:
      - description: ID of course to get grades for
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A grade distribution array for the course
          schema:
            items:
              type: integer
            type: array
...
  /grades/overall:
    get:
      description: '"Returns the overall grade distribution"'
      operationId: gradeAggregationOverall
      parameters:
      - description: The course's subject prefix
        in: query
        name: prefix
        type: string
      - description: The course's official number
        in: query
        name: number
        type: string
      - description: The professor's first name
        in: query
        name: first_name
        type: string
      - description: The professors's last name
        in: query
        name: last_name
        type: string
      - description: The number of the section
        in: query
        name: section_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A grade distribution array
          schema:
            items:
              type: integer
            type: array
  /grades/semester:
    get:
      description: '"Returns grade distributions aggregated by semester"'
      operationId: gradeAggregationBySemester
      parameters:
      - description: The course's subject prefix
        in: query
        name: prefix
        type: string
      - description: The course's official number
        in: query
        name: number
        type: string
      - description: The professor's first name
        in: query
        name: first_name
        type: string
      - description: The professors's last name
        in: query
        name: last_name
        type: string
      - description: The number of the section
        in: query
        name: section_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: An array of grade distributions for each semester included
          schema:
            items:
              $ref: '#/definitions/responses.GradeResponse'
            type: array
  /grades/semester/sectionType:
    get:
      description: '"Returns the grade distributions aggregated by semester and broken
        down into section type"'
      operationId: gradeAggregationSectionType
      parameters:
      - description: The course's subject prefix
        in: query
        name: prefix
        type: string
      - description: The course's official number
        in: query
        name: number
        type: string
      - description: The professor's first name
        in: query
        name: first_name
        type: string
      - description: The professors's last name
        in: query
        name: last_name
        type: string
      - description: The number of the section
        in: query
        name: section_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: An array of grade distributions for each section type for each
            semester included
          schema:
            items:
              $ref: '#/definitions/responses.SectionGradeResponse'
            type: array
...
  /professor/{id}/grades:
    get:
      description: '"Returns the overall grade distribution for a professor"'
      operationId: GradesByProfessorID
      parameters:
      - description: ID of professor to get grades for
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A grade distribution array for the professor
          schema:
            items:
              type: integer
            type: array
...
  /section/{id}/grades:
    get:
      description: '"Returns the overall grade distribution for a section"'
      operationId: GradesBySectionID
      parameters:
      - description: ID of section to get grades for
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A grade distribution array for the section
          schema:
            items:
              type: integer
            type: array
...
schemes:
- http
- https
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: x-api-key
    type: apiKey
  storage_key:
    in: header
    name: x-storage-key
    type: apiKey
swagger: "2.0"
...